# openapi.yaml
openapi: 3.1.0
info:
  title: Batchwise API
  version: 1.0.0
  description: Multi-tenant brewery management API.
servers:
  - url: http://localhost:8080
    description: Local development
  - url: https://api.batchwise.example
    description: Production (configurable)
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    RegisterRequest:
      type: object
      required: [email, password, display_name]
      additionalProperties: false
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 12, maxLength: 128 }
        display_name: { type: string, minLength: 1, maxLength: 100 }
        tenant_name: { type: string, minLength: 1, maxLength: 100 }
        country: { type: string, minLength: 2, maxLength: 2 }
        region: { type: string, maxLength: 100 }
    LoginRequest:
      type: object
      required: [email, password]
      additionalProperties: false
      properties:
        email: { type: string, format: email }
        password: { type: string }
...
    AuthResponse:
      type: object
      required: [user_id, tenant_id, email, display_name, role, access_token, refresh_token, token_type, expires_in]
      properties:
        user_id: { type: string, format: uuid }
        tenant_id: { type: string, format: uuid }
        email: { type: string, format: email }
        display_name: { type: string }
        role: { type: string, enum: [owner, manager, brewer, sales, viewer], description: "What the user may do; see docs/remediation-plan.md, Phase 15." }
        access_token: { type: string }
        refresh_token: { type: string }
        token_type: { type: string, enum: [Bearer] }
        expires_in: { type: integer, description: "Access token TTL in seconds" }
...
  responses:
    BadRequest:
      description: Validation failure or malformed JSON
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    Unauthorized:
      description: Missing or invalid token
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
...
    Conflict:
      description: Duplicate resource
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
...
paths:
  /api/v1/auth/register:
    post:
      operationId: register
      summary: Register a new user and tenant
      security: []
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RegisterRequest" }
      responses:
        "201":
          description: User and tenant created
          headers:
            Location:
              schema: { type: string }
              description: URL of the new user resource
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthResponse" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "409": { $ref: "#/components/responses/Conflict" }
  /api/v1/auth/login:
    post:
      operationId: login
      summary: Log in with email and password
      security: []
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LoginRequest" }
      responses:
        "200":
          description: Successful login
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403":
          description: Account inactive
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
  /api/v1/auth/refresh:
    post:
      operationId: refresh
      summary: Exchange a refresh token for a new token pair
      security: []
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RefreshRequest" }
      responses:
        "200":
          description: New token pair issued
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403":
          description: Account inactive
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
...
  /api/v1/members:
    get:
      operationId: listMembers
      summary: List the tenant's members (Owners and Managers)
      tags: [members]
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/MemberList" } } } }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /api/v1/members/{id}:
    patch:
      operationId: patchMember
      summary: Change a member's role or active flag
      description: Owners manage every member; Managers only Brewer, Sales and Viewer members, and only to one of those roles. A tenant must keep an active Owner (422 `last_owner`). Changes are audited and apply from the member's next request.
      tags: [members]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/PatchMemberRequest" } } }
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/Member" } } } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "422": { $ref: "#/components/responses/UnprocessableEntity" }
...
  /api/v1/inventory:
    get:
      operationId: listIngredients
      summary: List ingredient lots
      tags: [inventory]
      parameters:
        - { name: type, in: query, schema: { type: string } }
        - { name: name, in: query, schema: { type: string } }
        - { name: lot_number, in: query, schema: { type: string } }
        - { name: expiring_before, in: query, schema: { type: string, format: date } }
        - { name: expiring_within_days, in: query, schema: { type: integer } }
        - { name: out_of_stock, in: query, schema: { type: boolean } }
        - { name: sort, in: query, schema: { type: string } }
        - { name: page, in: query, schema: { type: integer } }
        - { name: page_size, in: query, schema: { type: integer } }
      responses:
        "200":
          description: Paginated ingredient lots
          content: { application/json: { schema: { allOf: [{ $ref: "#/components/schemas/Pagination" }, { properties: { items: { items: { $ref: "#/components/schemas/Ingredient" } } } }] } } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    post:
      operationId: createIngredient
      summary: Create an ingredient lot
      tags: [inventory]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/CreateIngredientRequest" } } }
      responses:
        "201":
          description: Ingredient lot created
          headers:
            Location: { schema: { type: string } }
          content: { application/json: { schema: { $ref: "#/components/schemas/Ingredient" } } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
...
  /api/v1/inventory/{id}:
    get:
      operationId: getIngredient
      summary: Get an ingredient lot by ID
      tags: [inventory]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200": { description: Ingredient, content: { application/json: { schema: { $ref: "#/components/schemas/Ingredient" } } } }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
    put:
      operationId: replaceIngredient
      summary: Replace an ingredient lot
      tags: [inventory]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/CreateIngredientRequest" } } }
      responses:
        "200": { description: Updated, content: { application/json: { schema: { $ref: "#/components/schemas/Ingredient" } } } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      operationId: patchIngredient
      summary: Partially update an ingredient lot
      tags: [inventory]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/PatchIngredientRequest" } } }
      responses:
        "200": { description: Updated, content: { application/json: { schema: { $ref: "#/components/schemas/Ingredient" } } } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      operationId: deleteIngredient
      summary: Delete an ingredient lot
      tags: [inventory]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "204": { description: Deleted }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
...
  /api/v1/recipes:
    get:
      operationId: listRecipes
      summary: List recipes
      tags: [recipes]
      parameters:
        - { name: name, in: query, schema: { type: string } }
        - { name: type, in: query, schema: { type: string } }
        - { name: style_id, in: query, schema: { type: string, format: uuid } }
        - { name: sort, in: query, schema: { type: string, default: "-created_at" } }
        - { name: page, in: query, schema: { type: integer, default: 1 } }
        - { name: page_size, in: query, schema: { type: integer, default: 20 } }
      responses:
        "200": { description: Paginated recipes, content: { application/json: { schema: { $ref: "#/components/schemas/RecipePage" } } } }
        "401": { $ref: "#/components/responses/Unauthorized" }
    post:
      operationId: createRecipe
      summary: Create a new recipe
      tags: [recipes]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/CreateRecipeRequest" } } }
      responses:
        "201": { description: Created, content: { application/json: { schema: { $ref: "#/components/schemas/RecipeWithIngredients" } } } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/UnprocessableEntity" }
...
  /api/v1/recipes/{id}:
    get:
      operationId: getRecipe
      summary: Get a recipe with all ingredients
      tags: [recipes]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200": { description: Recipe, content: { application/json: { schema: { $ref: "#/components/schemas/RecipeWithIngredients" } } } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
    put:
      operationId: replaceRecipe
      summary: Fully replace a recipe and all its ingredients
      tags: [recipes]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/CreateRecipeRequest" } } }
      responses:
        "200": { description: Updated, content: { application/json: { schema: { $ref: "#/components/schemas/RecipeWithIngredients" } } } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/UnprocessableEntity" }
    patch:
      operationId: patchRecipe
      summary: Partially update a recipe
      tags: [recipes]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/PatchRecipeRequest" } } }
      responses:
        "200": { description: Updated, content: { application/json: { schema: { $ref: "#/components/schemas/RecipeWithIngredients" } } } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      operationId: deleteRecipe
      summary: Delete a recipe
      tags: [recipes]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "204": { description: Deleted }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "422": { $ref: "#/components/responses/UnprocessableEntity" }
  /api/v1/batches:
    get:
      operationId: listBatches
      summary: List batches
      tags: [batches]
      parameters:
        - { name: status, in: query, schema: { type: string } }
        - { name: recipe_id, in: query, schema: { type: string, format: uuid } }
        - { name: brew_date_from, in: query, schema: { type: string, format: date } }
        - { name: brew_date_to, in: query, schema: { type: string, format: date } }
        - { name: page, in: query, schema: { type: integer } }
        - { name: page_size, in: query, schema: { type: integer } }
        - { name: sort, in: query, schema: { type: string } }
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/PaginatedBatches" } } } }
        "401": { $ref: "#/components/responses/Unauthorized" }
    post:
      operationId: createBatch
      summary: Create a batch
      tags: [batches]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/CreateBatchRequest" } } }
      responses:
        "201": { description: Created, content: { application/json: { schema: { $ref: "#/components/schemas/CreateBatchResponse" } } } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
  /api/v1/batches/{id}:
    get:
      operationId: getBatch
      summary: Get a batch
      tags: [batches]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/Batch" } } } }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
    put:
      operationId: replaceBatch
      summary: Replace a batch
      tags: [batches]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/UpdateBatchRequest" } } }
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/Batch" } } } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "422": { $ref: "#/components/responses/UnprocessableEntity" }
    patch:
      operationId: patchBatch
      summary: Partially update a batch
      tags: [batches]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/UpdateBatchRequest" } } }
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/Batch" } } } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "422": { $ref: "#/components/responses/UnprocessableEntity" }
    delete:
      operationId: deleteBatch
      summary: Delete a batch
      description: Only planned batches can be deleted. Once brewing has started the batch is a production record and stays, even if cancelled.
      tags: [batches]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "204": { description: Deleted }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "422": { $ref: "#/components/responses/UnprocessableEntity" }
...
security:
  - bearerAuth: []